+++ src/page/AdminPage/AdminLayout/AdminLayout.scss
@@ -1,5 +1,5 @@
-.admin-mobile-header_nav a:hover,
-.admin-mobile-header_nav .adminLinkActive {
+.admin-layout_nav a:hover,
+.admin-layout_nav .adminLinkActive {
   font-weight: 700 !important;
   color: #e77605 !important;
   border-bottom: solid 2px #e77605;
