+++ src/style/main.scss
@@ -20,7 +20,7 @@
 @import "../component/Footer/Footer.scss";
 @import "../component/Navbar/Navbar.scss";
 @import "../component/ArticleCard/ArticleCard.scss";
-@import "../component/AdminMobileHeader/AdminMobileHeader.scss";
+@import "../page/AdminPage/AdminLayout/AdminLayout.scss";
 @import "../page/AdminPage/adminInfo.scss";
 @import "../page/AdminPage/AdminBackground.scss";
 @import "../page/BlogPage/BlogHome.scss";
